backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
python-jose[cryptography]
passlib[bcrypt]
bcrypt==4.0.1
python-dotenv
pydantic[email]
pytest
httpx
python-multipart
psycopg2-binary
